+++ Api-Module/src/main/kotlin/com/bamyanggang/apimodule/domain/strongpoint/application/dto/CreateStrongPoint.kt
@@ -1,4 +1,4 @@
-package com.bamyanggang.apimodule.domain.experience.application.dto
+package com.bamyanggang.apimodule.domain.strongpoint.application.dto
 
 import java.util.*
