+++ Api-Module/src/main/kotlin/com/bamyanggang/apimodule/domain/strongpoint/presentation/StrongPointApi.kt
@@ -1,4 +1,4 @@
-package com.bamyanggang.apimodule.domain.experience.presentation
+package com.bamyanggang.apimodule.domain.strongpoint.presentation
 
 object StrongPointApi {
     const val BASE_URL = "/api/strong-points"
